split codec into more files
